improve rating DB update strategy
This introduces two changes:

1. Instead of iterating gradient ascent on the rating DB until it
stabilizes, we pick a number of iterations to do (more or less
arbitrarily) and stop after that many.
2. We make partial ratings estimates visible from the recommendation UI
more frequently.

This should address two issues we've been seeing and which I haven't
satisfactorily diagnosed:

1. Ratings inference sometimes appears to loop indefinitely. (I recently
killed a process with 290 hours of CPU time that had not yet chosen an
initial rating DB.)
2. It is near impossible to introspect into the system when something
has gone wrong.