make objective function game-count-invariant
We're doing gradient ascent. Gradient ascent works best if the magnitude
of the gradient is relatively predictable, so that you can choose an
appropriate learning rate.

Previously, our objective function added up the log-likelihood of all
the (relevant) games in the database. This sucks, because now the
gradient grows more quickly if there are more games contributing to the
sum than if there are fewer.

To even out this effect, we now divide the previous objective function
by the number of games that contributed to it.