bound the maximum rating change per step
Ratings inference works by guessing a rating for everybody and gradually
nudging it towards ratings that make the observations more likely.
Generally if the guess is a little bit off this isn't a big deal, as it
will get nudged in the right direction.

However, if the guess is off by a lot, that does become a problem. If
you get far enough off that the probability of your games gets rounded
off to zero, then all of a sudden your rating is no longer changing:
there is no gradient, because small changes to your ratings will keep
you in the "all probabilities fall below the rounding error" zone. That
sucks.

We've hit this several times, in several different ways. We've got fixes
for each, but it seems likely that we'll find other ways. So this commit
introduces an additional constraint on ratings inference aimed at
mitigating this problem. Namely, it introduces a maximum percentage
change that your rating can undergo in any single step of gradient
ascent. Currently, this is set to 50%, so if you are currently rated at
3.0, on the next step of ratings inference you will never go above 4.5
or below 2.0. Hopefully this should prevent wild swings where one player
goes from an undervalued rating of 0.01 to a wildly overvalued rating of
80; they'll still move in the right direction, but hopefully it will
take enough steps to get there that we will see an intermediate value
between 0.01 and 80 where they are slightly overvalued but not to the
point of zero-gradient.