prepare for team games
In this commit, we change the server so that it can accept teams (lists
of players) everywhere it accepted single players before. We change the
client to accommodate the backend API changes.

As a bonus, this fixes a bug that, though it was never reported and may
never have been triggered, is very real nevertheless: we now correctly
handle / and other URI characters in players' names correctly. The fix
is to call encodeURIComponent at appropriate moments, and should
probably be backported to mcmario itself at some point.