switch from Fixed E100 to Double
For Dr. Mario, we had lots of issues with rounding because we were
computing like degree-100 polynomials that had to end up with a number
between 0 and 1. We were adding and subtracting numbers with dozens of
digits and they were supposed to *almost* exactly cancel out. Double is
terrible for that.

But the calculations we're doing with Rocket League are much simpler --
the numbers we're adding together have roughly the same magnitude as
their sum -- so let's switch back to Double, which is faster.